Added sections and background in BAR skin.
If no sections, the behaviour is like before.

If there are sections, the bar’s background changes to a fancy version
of the section’s color.

If barEffects is false (default is true), the bar’s background outside
sections is coloured using the barBacgroundColor. When barEffects is
true, the default fancy color derived from the barColor is used.