Rollback of pull-request #90
To be honest I did not really like the solution. Also the barEffect is a bit misused here. I think not every skin has to support all available features. But you can feel free to fork skins and create your own versions easily. I've tried different solutions to get a more consistent solution for this skin but the conical gradient is not "cheap" to create which is the reason why I did not implement sections here. The background color is like the gradient itself always based on the bar color.