Stop disabled autoScale overriding properties
Fixes a bug where setting the autoScale can still override properties
(even when it is set to "false") because it was being set in the loop
through the HashMap. Consequently, it was possible for autoScale to
override explicit settings (such as tick spacing) even if it is disabled
in the builder.

This was especially happening when autoScale is triggered from setting
another property, such as angleRange.

This also fixes https://github.com/HanSolo/Medusa/issues/200

Signed-off-by: Kah Goh <villastar@yahoo.com.au>